More localization. Addition of the mentioned-in-Info.txt formatting function. This relies on LCMS's (a required project) built in printf support. LCMS must be a project in the same workspace in order for this to compile. This is true for the main project as well as the resource project. Found the bug in opening a PDF. One change in this project from revision 129 in the original project switched from using Double specifically to "Number" the base type for Double, Integer, etc. in J2SE. By mistake Double was switched to Integer in PDFObject. Switching it back fixed the issue.

+++ PDFRenderer/src/com/sun/pdfview/PDFObject.java
@@ -446,7 +446,7 @@ public class PDFObject
         }
         else if (type == NUMBER)
         {
-            return ((Integer)value).intValue();
+            return ((Double)value).intValue();
         }
         
         // wrong type
